+++ survey/views.py
@@ -1,0 +1,10 @@
+from rest_framework.generics import RetrieveAPIView
+
+from survey.models import Survey
+from survey.serializers import SurveySerializer
+
+
+class SurveyDetailView(RetrieveAPIView):
+    permission_classes = []
+    queryset = Survey.objects
+    serializer_class = SurveySerializer